interfaces added for each class
- GoogleGroupsConnection interface and its GoogleGroupsConnectionImpl implementing class

- GoogleGroupsService interface and its GoogleGroupsServiceImpl implementing class
